+++ src/main/java/org/folio/consortia/service/impl/UserTenantServiceImpl.java
@@ -310,6 +310,7 @@ public class UserTenantServiceImpl implements UserTenantService {
     UserTenantEntity entity = new UserTenantEntity();
     TenantEntity tenant = new TenantEntity();
     tenant.setId(userTenantDto.getTenantId());
+    tenant.setName(userTenantDto.getTenantName());
     tenant.setConsortiumId(consortiumId);
 
     if (Objects.nonNull(userTenantDto.getId())) {